Gave dtype to scale for bfloat support.

+++ rotary_embedding_torch/rotary_embedding_torch.py
@@ -106,7 +106,7 @@ class RotaryEmbedding(nn.Module):
         device, dtype, seq_len = q.device, q.dtype, q.shape[seq_dim]
         seq = torch.arange(seq_len, dtype = dtype, device = device)
         freqs = self.forward(lambda: seq, cache_key = f'freqs:{seq_len}')
-        scale = self.get_scale(lambda: seq, cache_key = f'scale:{seq_len}')
+        scale = self.get_scale(lambda: seq, cache_key = f'scale:{seq_len}').to(dtype)
         rotated_q = apply_rotary_emb(freqs, q, scale = scale)
         rotated_k = apply_rotary_emb(freqs, k, scale = scale ** -1)
         return rotated_q, rotated_k
